// src/app/models/application.model.ts
import { InternshipOffer } from "./internship-offer.model";

export interface Application {
  id: number; // Optional since it might be generated by the backend
  userId: number;
  internshipOffer: InternshipOffer; // Reference to InternshipOffer (optional if not always needed)
  fullName: string;
  email: string;
  phone: string;
  city: string;
  careerObjective: string;
  technologyProfile: string;
  skills: string[]; // List of skills
  experience: string;
  submissionDate: Date | string; // Can be Date or string depending on how backend returns it
  status: 'ACCEPTED' | 'WAITING' | 'DENIED';  // ✅ Ajout du statut ici !
}